add unit tests for CalculateInsurance method
